[BACKPORT 2025.1][#26986] YSQL: ASH: Rename wait_event_type "Network" to "RPCWait"
Summary:
Original commit: 34a34b881c12 / D43694

In all the occurrences of "Network" wait type, we are actually waiting for a RPC to complete and not only the network. So it is appropriate to rename wait_event_type "Network" to "RPCWait".

Jira: DB-16459

Test Plan: Only a rename. Testing is manual.

Reviewers: amitanand, asaha

Reviewed By: asaha

Subscribers: svc_phabricator, yql

Differential Revision: https://phorge.dev.yugabyte.com/D43944