[#25949] YSQL: Clean up TestPg15Regress
Summary:
Cleanup TestPg15Regress, which was temporarily added while stabilising the pg15 branch. These tests are duplicate subsets of other regression tests and are no longer required.

There are 15 YB_TODOs in this test file.
  - 13 of these are related to test tracking. All these are obsolete given that all the tests are already being tracked via detective.
  - The remaining two (described below) are non-reproducible anymore and hence, presumed to be fixed over the course of pg15 stabilisation.

One of the YB_TODO is related to failure (segmentation fault AFAIR) on changing the order of the tests.
```
-- Joins (YB_TODO: if I move it below pushdown test, the test fails)
```

The other YB_TODO is related to the difference in JOIN plan.
```
-- YB_TODO: pg15 used merge join, whereas hash join is expected.
```

As mentioned above, both of these are no longer reproducible.
Jira: DB-15268

Test Plan:
jenkins: compile-only

Close: #25949

Reviewers: telgersma

Reviewed By: telgersma

Subscribers: svc_phabricator, yql

Differential Revision: https://phorge.dev.yugabyte.com/D43704