[PLAT-17244][PLAT-17586]: Store upgrade details in audit table
Summary:
Added changes to store the finalizeRequired and ysqlMajorVersionUpgrade fields in the audit table.
YBM has moved from persistent disk to ephemeral volumes for YBA, which wipes out the metadata extracted from the DB package. So, if YBM still uses the precheck API, then YBA downloads the file and returns the result, but this causes a timeout in YBM. So, YBM will run the precheck/software upgrade API, and then they can extract this info.

Also, added the changes to make PG upgrade check work if auth is disabled in userIntent but enabled through gflags override.

Test Plan:
Tested manually that requested data is available in the audit table and the PG upgrade check works if auth is enabled using gflags override.

https://jenkins.dev.yugabyte.com/view/Dev/job/dev-itest-pipeline/1840/

Reviewers: daniel, amindrov, sanketh

Reviewed By: amindrov

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D43954