[#27155] DocDB: Protect changing format_version field by a kExternal autoflag
Summary:
e9aeaec84f71 Changed the `format_version` field of the `SnapshotInfoPB` from 2 to 3. This change is protected by the autoflag `enable_export_snapshot_using_relfilenode`, which was of type `kLocalVolatile`. However, as this `format_version` is persisted on disk as part of the backup, the autoflag should be `kExternal` because backups have an external scope (can be used for xCluster). Additionally, having the autoflag as `kLocalVolatile` caused restore failures when creating a backup during the monitoring phase of an upgrade and then restoring this backup after rolling back the upgrade.

Fixed the issue by making the autoflag `enable_export_snapshot_using_relfilenode` of class `kExternal`.

Additionally, this diff fixes an issue encountered when restoring a backup created during the upgrade window, and the restore happens after rolling back to the old version of ybdb. The error was that the GUC `yb_ignore_relfilenode_ids` is unrecognized by the old version, as this GUC has been introduced in the new version of the ybdb. To make this GUC setting backward compatible. Added a function that checks for the presence of the GUC before setting it.

**Upgrade/Downgrade safety**
This diff overrides the class of an existing autoflag. Therefore, we need to override the original class in `previous_auto_flags.json`. As the old value is already part of 2025.1, a backport is required to 2025.1 as well.

Jira: DB-16639

Test Plan:
Tested manually, by creating a backup in the monitoring phase and restoring it after rolling back the universe to the old version. Additionally, added the following backup upgrade test:

./yb_build.sh --cxx-test integration-tests_backup_upgrade-test --gtest_filter BackupUpgradeTest.TestRestoreBackupAfterRollback

Reviewers: hsunder, slingam

Reviewed By: hsunder, slingam

Subscribers: yql, ybase, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D43968