[BACKPORT 2025.1][#27035] DocDB: Fix interim pg_locks failure when one node is unavailable
Summary:
Original commit: ad8d14ccefb8c5fc9ae891e1d12ec0f9294503d4 / D43668
When a node goes down, master removes it from the list of live tservers after it fails to receive any heartbeat for a default of 60s - controlled by `FLAGS_tserver_unresponsive_timeout_ms`. Until then, the other tserver too treat this tserver as active.

While executing the pg_locks query, we fetch the list of status tablets on each tserver by querying the master. Post that, we try fetching the old transactions at all of the tservers considered live, and fail the query if we don't hear back from any tserver.

**Problem**
If a node goes down, the pg_locks query might fail for 60 secs post that.

**Solution**
Ideally, we could log and ignore this failure as we have a subsequent check where we fail the query if we haven't heard back a valid response for all of the status tablets. So it shouldn't be a problem even if a node goes down as the leaders switch to live nodes, and the pg_locks query can be served.
Jira: DB-16509

Test Plan:
Jenkins

./yb_build.sh --cxx-test='TEST_F(PgGetLockStatusTestRF3, TestPgLocksAfterTserverShutdown)'

Reviewers: yyan, rthallam, #db-approvers

Reviewed By: rthallam, #db-approvers

Subscribers: svc_phabricator, yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D43925