[#25892] YSQL: make dummy_seclabel in build_postgres.py
Summary:
Commit 57539dd6b797d99678c6f7c25296f16cb5b91dd2 adds dummy_seclabel as a
subdir to build for the default make.  It technically doesn't belong in
the default path, and the modification to add that there increases diff
with upstream postgres.

Instead, run make for dummy_seclabel in build_postgres.py.  From the
user perspective, it still builds by default with yb_build.sh, but the
intent is cleaner, and it can possibly be built when-needed in the
future.
Jira: DB-15199

Test Plan:
On Almalinux 8:

    ./yb_build.sh fastdebug --gcc11 --rebuild-postgres \
      --java-test TestPgRegressModulesDummySeclabel

Close: #25892

Reviewers: asaha

Reviewed By: asaha

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D43999