[#27152] docdb: tservers include their lease epoch in ysql lease refresh requests
Summary:
Currently tservers do not report what they think their lease epoch is when they refresh their ysql lease with the master leader. If the response from the master leader is lost the first time a tserver acquires a new lease, then the tserver will never restart its pg sessions and update its local copy of its lease epoch. This will prevent the tserver from hosting any DDLs until it is restarted.

This diff adds what the tserver thinks its current lease epoch is to the lease refresh request. The master leader will set the new_lease bit and provide the acquired locks to bootstrap both if it is giving the tserver a new lease, and also if the tserver does not supply the correct lease epoch in its lease refresh request. This way regardless of how many lease refresh requests are lost when the communication is successful the tserver will bootstrap and restart PG appropriately.

This diff also removes the `needs_bootstrap` field from `RefreshYsqlLeaseRequestPB`. It is redundant with `current_lease_epoch` and removing it streamlines the logic. This removal was done unsafely because the `needs_bootstrap` field only went into the 2025.1 branch and it hasn't shipped to any customers yet. We plan to backport this change before 2025.1 is shipped to any customers.

**Upgrade/Rollback safety:**
This diff contains modifications to protos whose use is gated behind test flags in all shipped versions. We will backport this diff to 2025.1 before it ships.
Jira: DB-16636

Test Plan:
```
% ./yb_build.sh fastdebug --cxx-test object_lock-test --gtest_filter 'ExternalObjectLockTest.RefreshYsqlLease'
% ./yb_build.sh fastdebug --cxx-test master-test --gtest_filter 'MasterTest.RefreshYsqlLease'
```

Reviewers: amitanand, #db-approvers

Reviewed By: amitanand

Subscribers: svc_phabricator, yql, ybase, rthallam, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D43963