[BACKPORT 2025.1][#27239] YSQL: avoid one type of full catalog cache refresh
Summary:
Currently, `SetYsqlDBCatalogVersions` and `SetYsqlDBCatalogInvalMessages` are
each atomic (using the same `lock_`) but they are not atomic as a whole. So
after `SetYsqlDBCatalogVersions` has returned, the `lock_` is released. Then
when `SetYsqlDBCatalogInvalMessages` is called, the `lock_` is acquired again.
In analyzing a recent integration regression test run logs, I found that it is
possible that when the cluster is loaded, there are many PG backends eager to
call `GetTserverCatalogMessageLists`. If `GetTserverCatalogMessageLists` is called
after `SetYsqlDBCatalogVersions` and before `SetYsqlDBCatalogInvalMessages`, it
will only see the new catalog version but not its associated invalidation
messages.  As a result, this PG backend will think the invalidation messages are
not available for the new version, so it goes ahead to perform a full catalog
cache refresh.  I made a change to call `SetYsqlDBCatalogVersions` and
`SetYsqlDBCatalogInvalMessages` atomically to avoid this type of full catalog
cache refreshes.

Also changed an existing log to make it more verbose when a full catalog cache
refresh is performed.
Jira: DB-16724

Original commit: 1a6ab902c2a6b940d50727b2602d6c168a76c232 / D44045

Test Plan: ./yb_build.sh --cxx-test pg_catalog_version-test

Reviewers: #db-approvers, kfranz, sanketh, mihnea

Reviewed By: sanketh

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D44139