[BACKPORT 2025.1][#3109][#26707] YSQL: Fix use after free on postponed DDL operations
Summary:
There are some DDL operations that are executed after the transaction commits. One such example is the drop of all sequences of a database when we drop the database. With the introduction of transactional DDL via 83d3cca918ed / D41998, we've introduced a use-after-free bug on such postponed operations.

Example:
```
create database d;
drop database d;
```

With transactional ddl, there is a slight change in the flow of an auto-committed DDL statement which execute postponed operations. The below comparison is for the `DROP DATABASE` command.

Before:
1. PG starts its own transaction (No DocDB txn)
2. A separate DDL transaction is started upon getting a DDL (YBTxnDdlProcessUtility calls YBIncrementDdlNestingLevel)
3. DDL statement (DROP DATABASE here) is executed which stores postponed DDL op to drop all sequences. This operation is allocated in the **`PortalContext`** and stored in the `CurrentTransactionState` via the `YBSaveDdlHandle` call.
4. DDL transaction is auto-committed (YBTxnDdlProcessUtility calls YBDecrementDdlNestingLevel).
5. Postponed DDL operations are executed as part of the auto-commit.
6. The execution of the DDL statement is considered finished and the `PortalContext` is freed.
7. The PG transaction auto-commits

After:
1. PG starts its own transaction
2. DDL statement is executed. No separate DDL txn is created. As before, the drop database command allocates the postponed operations in the `PortalContext`,
3. Once the execution of the DDL statement is finished, the `PortalContext` is freed.
4. PG auto-commits and the DDL changes are also committed as part of it (CommitTransaction -> YBCCommitTransaction -> YBCommitTransactionContainingDDL). Postponed DDL operations are executed as part of the auto-commit.

The difference in the before-after comparison is that earlier, the postponed ddl ops were executed before `PortalContext` was freed (end of statement execution) and now they are executed after (at commit of txn). So the handles that we save for postponed ddl ops need to outlive the statement lifetime i.e. the PortalContext.

##### Fix
This revision makes the change to allocate the handles in `CurTransactionContext`, instead of the earlier `PortalContext`. Also created a helper to clear the ddl_transaction_state to reduce code duplication.
Jira: DB-1839, DB-16083

Original commit: 85d404556138d7c488ab7b5c092d1c0405d681b1 / D43509

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestDdlTransactionBlocks'

Reviewers: myang, pjain, patnaik.balivada

Reviewed By: myang, patnaik.balivada

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D44102