[#26284] YSQL: Check for Postgres interrupts in PgGate
Summary:
It's been a known issue for a long time that Yugabyte does not respond
to interrupts promptly. Statement timeout mentioned in the ticket is
one case, but there are others, such as cancel requests (i.e. Ctrl-C
in the ysqlsh command line), shutdown requests were not handled while
Yugabyte was waiting for a response from DocDB.

Postgres provides CHECK_FOR_INTERRUPTS macro, which conventionally
called during long operations, most commonly in loops. However, that
macro is not available in PgGate's C++ code, and has typical for
Postgres limitations, most importantly, lack of thread safety.

This diff inroduces a function wrapper for the macro and passes it into
the PgGate as a callback. In PgGate a function is added, which performs
basic sanity checks, like if the function is called in the main thread,
and then calls the callback. That creates a framework, which enables
checks for Postgres interrupts from PgGate code.

The new framework is used in PgGate communications code, where delayed
reaction to interrupts is most noticeable. Before the change there
was a Get command which blocked until response is available, or error
occurs. This diff adds a loop before it, which calls WaitFor with short
timeout and checks for interrupts if timeout elapses before response is
available. The WaitFor is also introduced by this diff.
Jira: DB-15630

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgTimeout#testCancel'

Reviewers: dmitry, sergei, kramanathan

Reviewed By: kramanathan

Subscribers: jason, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43319