[#27297] YSQL: add multirange catalog entries
Summary:
Add catalog entries for the lsm access method for multirange types (see
upstream postgres commit 6df7a9698bb036610c1e8c6d375e1be38cb26d5f).
This mainly duplicates catalog entries for btree+multirange to
lsm+multirange.  This is similar to the duplication that has already
been done for btree+range to lsm+range.  There is a catalog pg_range
that has only btree+multirange, and that is left alone.

With these entries, creating an lsm index on multirange column is one
step closer to being supported.  Before, it fails due to lack of catalog
entries:

    ERROR:  data type nummultirange has no default operator class for access method "lsm"
    HINT:  You must specify an operator class for the index or define a default operator class for the data type.

Now, it fails due to YB not supporting the type as a key (see
yb_type.c YbTypeEntityTable):

    ERROR:  INDEX on column of type 'NUMMULTIRANGE' not yet supported

Port the remaining queries in regress test multirangetypes.  Index
creation fails, but keep the queries related to that even if they don't
exercise the index since, if those are removed, there aren't many
queries left to test.

Add a new YSQL migration upgrade snapshot for testing because adding a
migration script is currently not possible with this kind of catalog
change.  This should be fine as long as this change goes in to 2025.1.0.
Upgrading on the 2.25.x line may miss these catalog entries, but that
isn't the end of the world because these entries aren't on the critical
path.

Jira: DB-16793

Test Plan:
On Almalinux 8:

    ./yb_build.sh fastdebug --gcc11 --java-test TestPgRegressPgTypesUDT

Close: #27297

Reviewers: telgersma

Reviewed By: telgersma

Subscribers: kfranz, yql

Differential Revision: https://phorge.dev.yugabyte.com/D42128