[PLAT-17549] Don't mark nodes as To Be Removed if Full move fails to create new vms
Summary:
Added an ability to do full automatic rollback of edit universe when no new nodes were created.
In practice this usually means that there is a problem with new intended configuration of node (for example incompatible device type + instance type) so there is no sense in retrying.
So for this particular case we completely revert universe state so that other tasks could be run.
If at least one node was created successfully, this rollback is not applied.

One missing peace is we need to inform user that universe state is clean because of the auto rollback and customer is free to run any tasks - this would be done in a separate ticket

Test Plan:
1) Create a new universe
2) Add a piece of code to fail during node creation
3) Verify that universe task is rolled back, new tasks are available, no retry button in UI
--
4) Modify code to fail only for specific node
5) Verify that task failed and no rollback was performed

Reviewers: nsingh, anijhawan

Reviewed By: nsingh, anijhawan

Subscribers: svc_phabricator, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D44136