[BACKPORT 2025.1][#26626] Docdb: Fix clone database timeout to account for multi-region deployment
Summary:
Original commit: b70783f03260b9ccbacb1018c5064a7cfac0eec9 / D42968
In a multi-region cluster with a database that has 100s of schema objects like tables and sequences, it is expected that cloning the database will take some time due to the network roundtrips from tserver that is executing the dump script and master leader to clone the pg schema objects. That's why we introduced the flag `ysql_clone_pg_schema_rpc_timeout_ms` which is the timeout for cloning pg schema object. It is set to 10 minutes by default.
However, when executing the command `CREATE DATABASE clone_db TEMPLATE db1`, the command is timing out only after 2 minutes.

This results in the end user seeing a timeout when creating the clone database although the clone workflow might succeed under the hood (the clone workflow is asynchronous).
It turns out that `CREATE DATABASE` command has its own timeout that is specified in the function `CreateDatabaseDeadline()`. The timeout of the command is set to `yb_client_admin_operation_timeout_sec` which is 2 minutes.

The diff fixes the issue by overriding this timeout in case of clone to `ysql_clone_pg_schema_rpc_timeout_ms`. This accounts for the multi-region deployments and won't timeout the CREATE command for the end user.
Jira: DB-16001

Test Plan: ./yb_build.sh --cxx-test integration-tests_minicluster-snapshot-test --gtest_filter PgCloneTest.CloneYsqlDbTimeout

Reviewers: #db-approvers, asrivastava

Reviewed By: asrivastava

Subscribers: slingam, ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D44163