[PLAT-16470] Use long to store query call and row count
Summary:
**Context**
We've recently hit a case where the total query call count exceeded the range of a Java int.
This interger overflow caused the call stat (and related derived stats) to show up as negative numbers.

**Change**
Since we may overflow a Java int, this diff updates the data type to a long.

Test Plan:
As a sanity check, run some heavy workflow and ensure the slow query stats appear as expected.
Then verify that we don't overflow when the total call count or row count exceed the range of a Java int.
{F359070}

Reviewers: asharma, amalyshev

Reviewed By: asharma

Differential Revision: https://phorge.dev.yugabyte.com/D44246