[#27343] Docdb: Table Locks. Do not use auto flags to enable ysql lease mechanism
Summary:
Starting a TServer without enabling the ysql operation lease mechanism, and
enabling it later; can cause the TServer to restart Postgres due to the TServer
receiving a new lease.

This can be disrupting to the customers as their existing connections would all
be disconnected at about the same time.

It would be better to start the lease mechanism at tserver start by using a normal
gflag instead of an auto flag. Each TServer's lease is independent of other TServers;
and it only requires that the master is aware of the lease mechanism.

We do assume that the masters are upgraded first; before upgrading TServers.
Jira: DB-16852

Test Plan:
Jenkins
Upgrade tests

Reviewers: hsunder, rthallam, bkolagani

Reviewed By: rthallam

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D44257