[#27374] YSQL: Use highest bit mask for custom GUC stickiness with Connection Manager
Summary:
Connection Manager utilises bit-masking to denote stickiness for custom GUC variables. With regular PG15 merges that happen to introduce newer bit-mask flags, it is frequently missed to increment the custom stickiness flag to the next bit-mask value. This diff changes this mask value that connection manager uses to be the highest bit-mask value possible under the scope of an integer to avoid the overlooking of such issues.

Jira: DB-16889

Test Plan: Manual checks to ensure that setting GUC_ALLOW_IN_PARALLEL variables do not make the connection sticky with connection manager

Reviewers: skumar, arpit.saxena, mkumar

Reviewed By: arpit.saxena

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D44262