[#18728] YSQL: Fix flaky TestYsqlMetrics.testExplainMaxMemory
Summary:
`testExplainMaxMemory` has had a history of being flaky.
D29822 / 6b049039119 addressed it by comparing the median memory of the first n runs to the median memory
of the last n runs. This worked for a while, but then became flaky again, failing with:

```
TestYsqlMetrics.testExplainMaxMemory:538 Expected median memory to be consistent between first 10 runs and last 10 runs. Got measurements [332, 332, 332, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 364, 336, 332, 336, 332, 336, 332, 336, 332, 368, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 396, 336, 364, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332, 336, 332] expected:<332> but was:<336>
```

When the test was rewritten, the memory values were clustered around one value. Unrelated code changes have slowly shifted the distribution of memory, so that 332kB and 336kB are basically equally likely.
```
>>> from collections import Counter
>>> Counter(memories[:10])
Counter({332: 7, 336: 3})
>>> Counter(memories[-10:])
Counter({336: 5, 332: 5})
>>> Counter(memories)
Counter({332: 49, 336: 47, 364: 2, 368: 1, 396: 1})
```
The median is no longer a good tool to use for this, because the median is equally likely to be 332
as it is to be 336.

Instead, switch the approach to use the average, and compare the average with a tolerance. Since the first few runs are setting up caches,
ignore the first N_WARMUP_RUNS. This change is necessary because a median is more immune to outliers than a mean is.

Additionally, since this test takes a long time to run in sanitizers but just exercises the same
code over and over, reduce the number of iterations in sanitizer builds from 105 to 35.

Backport-through 2024.1.

This test is failing on 2.20, but for a different reason: `java.lang.AssertionError: Too many runs had an unexpected max memory.` so it requires a different fix.
Jira: DB-7624

Test Plan:
```
./yb_build.sh --java-test 'TestYsqlMetrics#testExplainMaxMemory'
```

Ran 100 times on both mac and linux.

Reviewers: jason, #db-approvers

Reviewed By: jason, #db-approvers

Subscribers: svc_phabricator, jason, smishra

Differential Revision: https://phorge.dev.yugabyte.com/D44112