[#26561] YSQL: Detoast values before inserting into DocDB table
Summary:
Temp tables (which use PG storage) may contain toasted (i.e. compressed) values.
When we `INSERT INTO permanent_table SELECT FROM temp_table`, we do not detoast
these values, leading to an error, usually `Unexpected value type: kCollString`.

This diff detoasts toasted values that are being inserted/updated to YB storage (DocDB) because DocDB currently does not support toasted values at all. This means that, if in the future, YB storage does support toast, depending on how it is implemented (e.g. if YB is able to support a mix of toast/no-toast), we may want to remove this workaround.

We implement detoasting by modifying the `YbDatumToX` functions that are used when converting a `Datum` to our DocDB data type. There are 3 of these:
1. `YbDatumToBinary`: used for 103 varlena PG types such as arrays & ranges
2. `YbDatumToDocdb`: miscellaneous conversion function used for 40 PG types including `JSON` & `TSVECTOR`. Supports both varlena and fixed-length types.
3. `YbDatumToText`: used only for the `TEXT` type

This means we are calling `PG_DETOAST_DATUM` on every variable-length attribute that we are sending to DocDB. This should not have a significant performance overhead because `pg_detoast_datum` checks the varlena header (a cheap operation) and returns early if it is not toasted.

Note that vanilla PG never detoasts on the insert/update/delete path--even if a table has a column with toast disabled, when a toasted attribute is inserted into this table it just keeps it toasted. See the `heap_insert` and `heap_update` for more details--we do not use these functions for DocDB tables.
Jira: DB-15929

Test Plan:
```
./yb_build.sh release --cxx-test pg_temp_toast-test
```

Reviewers: jason

Reviewed By: jason

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42935